+++ src/cli/dcos_docker/__init__.py
@@ -4,7 +4,10 @@ A CLI for controlling DC/OS clusters on Docker.
 
 import click
 
-from cli.common.commands import list_clusters_command_factory
+from cli.common.commands import (
+    download_artifact,
+    list_clusters_command_factory,
+)
 
 from .commands._common import existing_cluster_ids
 from .commands.create import create
@@ -34,6 +37,7 @@ dcos_docker.add_command(destroy)
 dcos_docker.add_command(destroy_list)
 dcos_docker.add_command(destroy_mac_network)
 dcos_docker.add_command(doctor)
+dcos_docker.add_command(download_artifact)
 dcos_docker.add_command(inspect_cluster)
 dcos_docker.add_command(LIST_CLUSTERS)
 dcos_docker.add_command(run)